Fix inductor `end` of slice.
This PR fixes checks for the `end` integer when slicing a tensor. The problem was that
slicing dimension `dim` with `(start, end, step)` created (among others) the guard: `end <
size[dim]`.

Although, that doesn't generate incorrect results, a tighter guard would be: `end <
size[dim] + 1`, since `end` is not included in the slice.

ghstack-source-id: faa49e8a6de35de706d96b6c7fa1802fe7e769ee
Pull Request resolved: https://github.com/pytorch/pytorch/pull/106208