[custom_ops] extend impl_abstract to work with existing torch.library ops
This PR extends impl_abstract to work with existing
torch.library/TORCH_LIBRARY ops.

There's a question of what to do if the user calls impl_abstract
and the op already has a registration for:
- DispatchKey::Meta. We raise.
- DispatchKey::CompositeImplicitAutograd. We raise.
- DispatchKey::CompositeExplicitAutograd. To be pragmatic, we don't
raise, since the user's CompositeExplicitAutograd might work for all
other backends but Meta.

Test Plan:
- new tests

ghstack-source-id: d637aa7deb279890c33f7e5e5d9208e3a071f912
Pull Request resolved: https://github.com/pytorch/pytorch/pull/106088